[BUGFIX] Critical Invalid Cache Problem
The problem occurs when some language packs are installed
and the models are reflected. There are situations where the classes
(e.g. Country) are reflected and the Static Info Tables cache
is not up-to-date.
Solution: drop the reflection cache in the rebuild process
of the Static Info Tables objects.

Thanks to Tim Lochmüller

Change-Id: Id9d9e7aa3c0ed8f1bcbf489fde914efc7b247ac0
Resolves: #82808
Reviewed-on: https://review.typo3.org/54685
Reviewed-by: Stanislas Rolland <typo3@sjbr.ca>
Tested-by: Stanislas Rolland <typo3@sjbr.ca>