[BUGFIX] Restore compatibility with PHP 7.0
As this version of the extension is marked as compatible with TYPO3 8.7,
it should also be compatible with the lowest PHP version required for
TYPO3 8.7, i.e., PHP 7.0.

Hence, there must not be any nullable return type declarations.

Also state compatibility with PHP 7.0.x up to 7.2.x in the `composer.json`.
(This also allows PhpStorm to automatically determine the PHP language
level.)

Resolves: #86710

Change-Id: I41a14355573e85f59155ab892f1373af9d345166
Reviewed-on: https://review.typo3.org/58665
Reviewed-by: Stanislas Rolland <typo3@sjbr.ca>
Tested-by: Stanislas Rolland <typo3@sjbr.ca>