[BUGFIX] Add column names when inserting static data
When the extension is used in a composer-based installation and another
language is configured as well, just as e.g., static_info_tables_de,
additional columns are automatically created by TYPO3 for the translations
and this prevents the static data from being populated. By explicitly
giving the names of columns the extension will populate, the DBMS properly
uses default values for undefined columns.

Resolves: #82132
Change-Id: I3908b516a5250948d75cc12711e8765d457d10c6
Reviewed-on: https://review.typo3.org/c/TYPO3CMS/Extensions/static_info_tables/+/60798
Tested-by: Stanislas Rolland <typo3@sjbr.ca>
Reviewed-by: Stanislas Rolland <typo3@sjbr.ca>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>