[TASK] Only run update script when extensions have changed
To improve the performance for Composer-based installations using
TYPO3 console or for continuous deployments, the update script
for rebuilding the static_info_tables will only run when
there is a change in the installed extensions.

The status of the installed Extensions is stored in the registry.
When the stored versions did not change, the updater will skip
processing.

Resolves: #88684
Change-Id: I53892cd727ea3336e0c60b0f9ca8c49cb3a1d5ca
Reviewed-on: https://review.typo3.org/c/TYPO3CMS/Extensions/static_info_tables/+/61224
Tested-by: Stanislas Rolland <typo32020@sjbr.ca>
Reviewed-by: Stanislas Rolland <typo32020@sjbr.ca>